Splitting LinuxContext.cpp into LinuxContext.cpp & LinuxApp.cpp
