+++ platform/linux/CMakeList.txt
@@ -874,6 +874,7 @@ SET( SOLAR2D_SOURCES
 add_executable( Solar2D
 	${SOLAR2D_SOURCES}
 	${CORONA_ROOT}/platform/linux/src/main.cpp
+	${CORONA_ROOT}/platform/linux/src/Rtt_LinuxApp.cpp
 	${CORONA_ROOT}/platform/linux/src/Rtt_LinuxContext.cpp
 	${CORONA_ROOT}/platform/linux/src/Rtt_LinuxAudioRecorder.cpp
 	${CORONA_ROOT}/platform/linux/src/Rtt_LinuxBitmap.cpp
@@ -918,6 +919,7 @@ add_executable( Solar2D
 add_executable( Solar2DSimulator
 	${SOLAR2D_SOURCES}
 	${CORONA_ROOT}/platform/linux/src/main.cpp
+	${CORONA_ROOT}/platform/linux/src/Rtt_LinuxApp.cpp
 	${CORONA_ROOT}/platform/linux/src/Rtt_LinuxContext.cpp
 	${CORONA_ROOT}/platform/linux/src/Rtt_LinuxAudioRecorder.cpp
 	${CORONA_ROOT}/platform/linux/src/Rtt_LinuxBitmap.cpp
