+++ platform/linux/src/Rtt_LinuxBuildDialog.cpp
@@ -12,6 +12,7 @@
 #include "Rtt_LinuxSimulatorServices.h"
 #include "Rtt_LinuxSimulatorView.h"
 #include "Rtt_LinuxBuildDialog.h"
+#include "Rtt_LinuxApp.h"
 #include "Rtt_SimulatorRecents.h"
 #include "Rtt_WebAppPackager.h"
 #include "Rtt_LinuxAppPackager.h"
