+++ platform/linux/src/Rtt_LinuxRuntimeErrorDialog.cpp
@@ -1,5 +1,5 @@
 #include "Rtt_LinuxRuntimeErrorDialog.h"
-#include "Rtt_LinuxContext.h"
+#include "Rtt_LinuxApp.h"
 #include <string.h>
 
 using namespace std;
